adding an implementation that saves 'save' and 'delete' operations to the database

+++ src/main/java/com/example/propertysearcherprojectbackend/service/PropertyService.java
@@ -1,5 +1,6 @@
 package com.example.propertysearcherprojectbackend.service;
 
+import com.example.propertysearcherprojectbackend.domain.AuditMessage;
 import com.example.propertysearcherprojectbackend.domain.Property;
 import com.example.propertysearcherprojectbackend.dto.PropertyDto;
 import com.example.propertysearcherprojectbackend.exceptions.PropertyNotFoundException;
@@ -8,6 +9,7 @@ import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,6 +17,7 @@ import java.util.Optional;
 @AllArgsConstructor
 public class PropertyService {
     private final ModelMapper modelMapper;
+    private final AuditMessageService auditMessageService;
     private final PropertyRepository propertyRepository;
 
     public List<Property> getAllProperties() {
@@ -29,12 +32,26 @@ public class PropertyService {
         }
     }
 
-    public void deletePropertyById(Long propertyId) {
-        propertyRepository.deleteById(propertyId);
+    public void deletePropertyById(Long propertyId) throws PropertyNotFoundException {
+        if (!propertyRepository.existsById(propertyId)) {
+            throw new PropertyNotFoundException("Property with id %s not found");
+        } else {
+            auditMessageService.saveAuditMessage(AuditMessage.builder()
+                    .message("Property deleted: property id " + propertyId)
+                    .createdAt(LocalDateTime.now())
+                    .build());
+            propertyRepository.deleteById(propertyId);
+        }
     }
 
     public Property saveProperty(PropertyDto propertyDto) {
-        return propertyRepository.save(modelMapper.map(propertyDto, Property.class));
+       Property property = propertyRepository.save(modelMapper.map(propertyDto, Property.class));
+                auditMessageService.saveAuditMessage(AuditMessage.builder()
+                .message("New property added: " + property.getPropertyType() + " " + property.getAddress())
+                .createdAt(LocalDateTime.now())
+                .build());
+
+        return property;
     }
 
     public Property updateProperty(PropertyDto propertyDto, Long propertyId) throws PropertyNotFoundException {
